refactor(auth): パスキー関連ファイルを整理する
