+++ src/features/auth/components/passkey/list-item.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react'
 import { css } from 'styled-system/css'
 
-import { authClient } from '../lib/auth-client'
-import { formatPasskeyCreatedAt } from '../lib/format-passkey-created-at'
-import { passkeyDisplayName } from '../lib/passkey-display-name'
-import { getPasskeyManageErrorMessage } from '../lib/passkey-messages'
-import { PasskeyDeleteDialog } from './passkey-delete-dialog'
-import { PasskeyRenameDialog } from './passkey-rename-dialog'
+import { authClient } from '../../lib/auth-client'
+import { passkeyDisplayName } from '../../lib/passkey.display-name'
+import { formatPasskeyCreatedAt } from '../../lib/passkey.format-created-at'
+import { getPasskeyManageErrorMessage } from '../../lib/passkey.messages'
+import { PasskeyDeleteDialog } from './delete-dialog'
+import { PasskeyRenameDialog } from './rename-dialog'
 
 export type ManagedPasskey = {
   readonly id: string
