+++ src/features/auth/components/passkey/sign-in.tsx
@@ -4,11 +4,14 @@ import { CircleAlert, KeyRound } from 'lucide-react'
 import { useEffect, useRef, useState } from 'react'
 import { css } from 'styled-system/css'
 
-import { StyledButton } from '../../../shared/components/styled-button'
-import { formSummary } from '../../../styles/form'
-import { authClient } from '../lib/auth-client'
-import { getPasskeySignInErrorMessage } from '../lib/passkey-messages'
-import { isConditionalMediationAvailable, isWebAuthnAvailable } from '../lib/webauthn-support'
+import { StyledButton } from '../../../../shared/components/styled-button'
+import { formSummary } from '../../../../styles/form'
+import { authClient } from '../../lib/auth-client'
+import { getPasskeySignInErrorMessage } from '../../lib/passkey.messages'
+import {
+  isConditionalMediationAvailable,
+  isWebAuthnAvailable
+} from '../../lib/passkey.webauthn-support'
 
 const passkeySignIn = css({
   display: 'flex',
